requirements.txt
neptune==1.10.2
scikit-learn==1.4.2
neptune-sklearn==2.1.3

pydantic-settings==2.2.1
rich==13.7.1
